Divided tasks into classes, saved them in packagesÑ.
