+++ src/main/java/edu/hw8/task2/ThreadPool.java
@@ -1,0 +1,7 @@
+package edu.hw8.task2;
+
+public interface ThreadPool extends AutoCloseable {
+    void start();
+
+    void execute(Runnable runnable);
+}
